Introduced toString(), parse() to RecordIdentifier and removed helper
